2022-11-1 commit

See merge request s07-final/S07P31C101!66

+++ backend/src/main/java/com/handfarm/backend/service/impl/DeviceServiceImpl.java
@@ -98,7 +98,7 @@ public class DeviceServiceImpl implements DeviceService {
     @Override
     public JsonObject deviceAutoControl(String deviceNo, DedviceAutoControlDto dto) {
         String control = dto.getControlName();
-        Integer value = dto.getControlValue();
+        Integer value = (Integer) dto.getControlValue();
         Optional<ControlEntity> controlEntity = controlRepository.findByControlName(control);
         Optional<DeviceEntity> deviceEntity = deviceRepository.findByDeviceNo(deviceNo);
         Optional<DeviceControlEntity> deviceControlEntity = deviceControlRepository.findByDeviceIdxAndControlIdx(deviceEntity.get(), controlEntity.get());
@@ -138,7 +138,7 @@ public class DeviceServiceImpl implements DeviceService {
     @Override
     public JsonObject deviceManual(String deviceNo, DedviceAutoControlDto dto) {
         String control = dto.getControlName();
-        Integer value = dto.getControlValue();
+        Integer value = (Integer) dto.getControlValue();
         Optional<ControlEntity> controlEntity = controlRepository.findByControlName(control);
         Optional<DeviceEntity> deviceEntity = deviceRepository.findByDeviceNo(deviceNo);
         Optional<DeviceControlEntity> deviceControlEntity = deviceControlRepository.findByDeviceIdxAndControlIdx(deviceEntity.get(), controlEntity.get());
